Fix and update the receipt and storage of information and other things from sensors

+++ include/sensor_functions.h
@@ -3,5 +3,5 @@
 #include <TroykaMQ.h>
 
 void initialization();
-int readHeight();
-int readIllumination();
+double readHeight();
+double readIllumination();
